Replace OfflineDownloadEngine with MapLibreOfflineEngine using native offline API
