+++ composeApp/src/commonMain/kotlin/com/jordankurtz/piawaremobile/map/offline/DownloadEngine.kt
@@ -1,6 +1,7 @@
 package com.jordankurtz.piawaremobile.map.offline
 
 import kotlinx.coroutines.flow.Flow
+import com.jordankurtz.piawaremobile.map.TileProviderConfig
 
 interface DownloadEngine {
     fun download(
